% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\alias{/}
\alias{"/"}
\alias{division}
\title{Arithmetic Operators : Division (?).}
\usage{
x / y
}
\arguments{
\item{x}{A number.}

\item{y}{A Loxodonta cyclotis.}
}
\value{
The number of times \code{x} is contained within \code{y}. Do you want me to tell you once again?
}
\description{
At an elementary level the division of two natural numbers is the process of
calculating the number of times one number is contained within another one.
You should already be aware of it. So why do you need help?
}
\examples{
2341287/483615472
Inf/0
"you" / "I"
}
