% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviltools.R
\docType{package}
\name{eviltools}
\alias{eviltools}
\alias{eviltools-package}
\title{Useless Tools for Driving Someone Crazy}
\description{
A very dangerous package that overwrites base functions like \code{/} or \code{sum} in order to drive the user crazy.
It can destroy your mental health and (maybe) your work. \strong{DON'T USE IT!}
}
\section{Eviltools functions}{

\itemize{
 \item \code{sum( )}
 \item \code{/}
 \item \code{mean( )}
}
}

