% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{Generic mean (?)}
\usage{
mean (x, type)
}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric/logical vectors and date, date-time and time interval objects.}

\item{type}{A number from 1 to 5 that identifies which type of mean the user wants to calculate.}
}
\value{
The arithmetic, weighted, geometric, armonic or arithmetic-geometric mean of the values in x is computed, as a numeric or complex vector of length one.
If x is not logical (coerced to numeric), numeric (including integer) or complex, \code{NA_real_} is returned, with a warning.
}
\description{
Generic interactive function for the arithmetic, weighted, geometric, armonic or arithmetic-geometric mean.
}
\examples{
x <- c(13:20, -31.2, 4321, -24)
mean(x, type = 1) # don't input the "type", R will ask you what you want anyways.
}
