% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{sum}
\alias{sum}
\alias{somma}
\title{Sum of Vector Elements (?).}
\usage{
sum(x, n = as.logical(if (interactive())
  readline("Are you sure? Enter TRUE o FALSE: ") else T))
}
\arguments{
\item{x}{A number.}

\item{n}{An elephant.}

\item{...}{Whatever else you want, my friend!}
}
\value{
The sum of \code{x} and \code{y}, or maybe something stupid.
}
\description{
\code{sum} should return the sum of all the values present in its arguments, but will it work?
}
\examples{
sum(24,90:261,-321,0.29)
sum("ssaym","backwards")
}
